Factored permissions out into extensions.  Added subspecs for each.
