+++ PermissionScope.podspec
@@ -10,7 +10,54 @@ Pod::Spec.new do |s|
 
   s.ios.deployment_target = '8.0'
 
-  s.source_files = 'PermissionScope/*.swift'
-
   s.requires_arc = true
+  
+  s.default_subspec = 'Core'
+  
+  s.subspec 'Core' do |sub|
+    sub.source_files = 'PermissionScope/*.swift'
+  end
+
+  s.subspec 'Bluetooth' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Bluetooth.swift'
+  end
+  
+  s.subspec 'Camera' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Camera.swift'
+  end
+  
+  s.subspec 'Contacts' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Contacts.swift'
+  end
+
+  s.subspec 'Events' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Events.swift'
+  end
+
+  s.subspec 'Location' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Location.swift'
+  end
+  
+  s.subspec 'Microphone' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Microphone.swift'
+  end
+
+  s.subspec 'Motion' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Motion.swift'
+  end
+  
+  s.subspec 'Notifications' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Notifications.swift'
+  end
+
+  s.subspec 'Photos' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Photos.swift'
+  end
+
+  s.subspec 'Reminders' do |sub|
+    sub.source_files = 'PermissionScope/Extensions/PermissionScope+Reminders.swift'
+  end
+  
 end
+
+
